Small bug fixes

+++ Melon/Optimizing/KiwiOptimizer.h
@@ -11,7 +11,7 @@ namespace Melon {
 
 		class KiwiOptimizer {
 		public:
-			static Boxx::List<Kiwi::Instruction> Optimize(const Boxx::List<OptimizerInstruction>& instructionList, const Boxx::UByte loops = 1);
+			static Boxx::List<Kiwi::Instruction> Optimize(const Boxx::List<OptimizerInstruction>& instructionList, const Boxx::UInt loops = 1);
 
 		private:
 			static Boxx::List<Boxx::Pair<Boxx::List<OptimizerInstruction>, Boxx::List<OptimizerInstruction>>> Split(const Boxx::List<OptimizerInstruction>& instructions);
